+++ app/javascript/constants/ui.js
@@ -1,0 +1,4 @@
+export const SEVERITY = {
+  SUCCESS: 'success',
+  ERROR: 'error',
+};
