+++ app/javascript/helpers/index.js
