+++ app/javascript/repositories/UsersRepository.js
@@ -1,29 +1,29 @@
-import { apiV1TasksPath, apiV1TaskPath } from '../../routes/ApiRoutes';
-import FetchHelper from '../utils/fetchHelper';
+import { apiV1UsersPath, apiV1UserPath } from 'routes/ApiRoutes';
+import FetchHelper from 'utils/fetchHelper';
 
 export default {
   index(params) {
-    const path = apiV1TasksPath();
+    const path = apiV1UsersPath();
     return FetchHelper.get(path, params);
   },
 
   show(id) {
-    const path = apiV1TaskPath(id);
+    const path = apiV1UserPath(id);
     return FetchHelper.get(path);
   },
 
   update(id, task = {}) {
-    const path = apiV1TaskPath(id);
+    const path = apiV1UserPath(id);
     return FetchHelper.put(path, task);
   },
 
   create(task = {}) {
-    const path = apiV1TasksPath();
+    const path = apiV1UsersPath();
     return FetchHelper.post(path, task);
   },
 
   destroy(id) {
-    const path = apiV1TaskPath(id);
+    const path = apiV1UserPath(id);
     return FetchHelper.delete(path);
   },
 };
