+++ app/javascript/utils/fetchHelper.js
