+++ config/webpack/custom.js
@@ -1,17 +1,17 @@
 const path = require('path');
 
-const reactFolder = path.resolve(__dirname, '../../app/javascript/react/');
+const reactFolder = path.resolve(__dirname, '../../app/javascript/');
 
 module.exports = {
-  entry: path.resolve(__dirname, '../../app/javascript/react/applications.js'),
+  entry: path.resolve(__dirname, '../../app/javascript/applications.js'),
   resolve: {
     alias: {
-      components: path.resolve(reactFolder, 'components'),
-      constants: path.resolve(reactFolder, 'constants.js'),
-      forms: path.resolve(reactFolder, 'forms'),
-      helpers: path.resolve(reactFolder, 'helpers'),
       repositories: path.resolve(reactFolder, 'repositories'),
+      components: path.resolve(reactFolder, 'components'),
       routes: path.resolve(reactFolder, 'repositories'),
+      constants: path.resolve(reactFolder, 'constants'),
+      helpers: path.resolve(reactFolder, 'helpers'),
+      forms: path.resolve(reactFolder, 'forms'),
     },
   },
 };
